feat(sample): update to nativescript 8 and react 18
