--- sample/.editorconfig
@@ -1,15 +1,0 @@
-root = true
-
-[*]
-end_of_line = lf
-insert_final_newline = true
-trim_trailing_whitespace = true
-charset = utf-8
-
-[*.json]
-indent_style = space
-indent_size = 2
-
-[*.ts]
-indent_style = space
-indent_size = 4
