+++ sample/src/app.css
@@ -10,11 +10,10 @@ of writing your own CSS rules. For a full list of class names in the theme
 refer to http://docs.nativescript.org/ui/theme.
 The imported CSS rules must precede all other types of rules.
 
-@example 
+@example
   @import "~@nativescript/theme/css/core.css";
   @import "~@nativescript/theme/css/blue.css";
 
 Jamie: I'll leave theming up to the consumer instead in RNS; it's too much
        of a black-box to me.
 */
-
