+++ sample/src/images/RNTester_Thumbnails/React_Native_LICENSE.txt
