+++ sample/src/images/RNTester_Thumbnails/bandaged.png
