+++ sample/src/images/RNTester_Thumbnails/call.png
