+++ sample/src/images/RNTester_Thumbnails/dislike.png
