+++ sample/src/images/RNTester_Thumbnails/fist.png
