+++ sample/src/images/RNTester_Thumbnails/flowers.png
