+++ sample/src/images/RNTester_Thumbnails/heart.png
