+++ sample/src/images/RNTester_Thumbnails/like.png
