+++ sample/src/images/RNTester_Thumbnails/liking.png
