+++ sample/src/images/RNTester_Thumbnails/poke.png
