+++ sample/src/images/RNTester_Thumbnails/superlike.png
