+++ sample/src/images/RNTester_Thumbnails/victory.png
