+++ sample/src/testComponents/AppContainer.tsx
@@ -2,14 +2,14 @@ import * as React from "react";
 import { StyleSheet } from "react-nativescript";
 import { Clock, ClockFC, Counter, MarqueeFC } from "./stateful";
 
-function AppContainer(){
+function AppContainer() {
     return (
         <stackLayout style={styles.stack}>
             <label>Hello world!</label>
-            <Clock/>
-            <ClockFC/>
-            <Counter/>
-            <MarqueeFC text="The quick brown fox jumps over the lazy dog."/>
+            <Clock />
+            <ClockFC />
+            <Counter />
+            <MarqueeFC text="The quick brown fox jumps over the lazy dog." />
         </stackLayout>
     );
 }
@@ -17,7 +17,7 @@ function AppContainer(){
 const styles = StyleSheet.create({
     stack: {
         backgroundColor: "gold",
-    }
+    },
 });
 
 export default AppContainer;
