+++ sample/src/testComponents/navigation.tsx
@@ -5,105 +5,131 @@ import { NSVElement, FrameAttributes, PageAttributes } from "react-nativescript"
 import { Page, Frame, NavigationButton } from "@nativescript/core";
 import { DockLayoutTest, FlexboxLayoutTest, AbsoluteLayoutTest } from "./layout";
 
-export class NestedHub extends React.Component<{
-    forwardedRef: React.RefObject<NSVElement<Page>>;
-} & PageAttributes, {}> {
+export class NestedHub extends React.Component<
+    {
+        forwardedRef: React.RefObject<NSVElement<Page>>;
+    } & PageAttributes,
+    {}
+> {
     render() {
         const { forwardedRef, ...rest } = this.props;
         const greenPageRef = React.createRef<NSVElement<Page>>();
         const redPageRef = React.createRef<NSVElement<Page>>();
-        return (<page ref={forwardedRef} actionBarHidden={false} {...rest}>
-            <actionBar title="Hub" className="action-bar" />
-            <stackLayout>
-                <button text={"Navigate to green page"} onTap={() => {
-                    const currentPage: Page = forwardedRef.current!.nativeView;
-                    currentPage.frame.navigate({
-                        create: () => {
-                            return greenPageRef.current.nativeView;
-                        }
-                    });
-                }} />
-            </stackLayout>
-
-            <PortalToPageWithActionBar forwardedRef={greenPageRef} actionBarTitle={"Green"} backgroundColor={"green"}>
+        return (
+            <page ref={forwardedRef} actionBarHidden={false} {...rest}>
+                <actionBar title="Hub" className="action-bar" />
                 <stackLayout>
-                    <label>You're viewing the green page!</label>
-                    <button text={"Navigate to red page"} onTap={() => {
-                        const currentPage: Page = greenPageRef.current!.nativeView;
-                        currentPage.frame.navigate({
-                            create: () => {
-                                return redPageRef.current.nativeView;
-                            }
-                        });
-                    }} />
+                    <button
+                        text={"Navigate to green page"}
+                        onTap={() => {
+                            const currentPage: Page = forwardedRef.current!.nativeView;
+                            currentPage.frame.navigate({
+                                create: () => {
+                                    return greenPageRef.current.nativeView;
+                                },
+                            });
+                        }}
+                    />
                 </stackLayout>
 
-            </PortalToPageWithActionBar>
+                <PortalToPageWithActionBar forwardedRef={greenPageRef} actionBarTitle={"Green"} backgroundColor={"green"}>
+                    <stackLayout>
+                        <label>You're viewing the green page!</label>
+                        <button
+                            text={"Navigate to red page"}
+                            onTap={() => {
+                                const currentPage: Page = greenPageRef.current!.nativeView;
+                                currentPage.frame.navigate({
+                                    create: () => {
+                                        return redPageRef.current.nativeView;
+                                    },
+                                });
+                            }}
+                        />
+                    </stackLayout>
+                </PortalToPageWithActionBar>
 
-            <PortalToPageWithActionBar forwardedRef={redPageRef} actionBarTitle={"Red"} backgroundColor={"red"}>
-                <stackLayout>
-                    <label>You're viewing the red page!</label>
-                </stackLayout>
-            </PortalToPageWithActionBar>
-        </page>);
+                <PortalToPageWithActionBar forwardedRef={redPageRef} actionBarTitle={"Red"} backgroundColor={"red"}>
+                    <stackLayout>
+                        <label>You're viewing the red page!</label>
+                    </stackLayout>
+                </PortalToPageWithActionBar>
+            </page>
+        );
     }
 }
-export class NestedModalTest extends React.Component<{
-    forwardedRef: React.RefObject<NSVElement<Page>>;
-} & PageAttributes, {}> {
+export class NestedModalTest extends React.Component<
+    {
+        forwardedRef: React.RefObject<NSVElement<Page>>;
+    } & PageAttributes,
+    {}
+> {
     render() {
         const { forwardedRef, ...rest } = this.props;
         const yellowPageRef = React.createRef<NSVElement<Page>>();
         const greenPageRef = React.createRef<NSVElement<Page>>();
-        return (<page ref={forwardedRef} actionBarHidden={false} {...rest}>
-            <actionBar title="Navigation Hub" className="action-bar" />
-            <stackLayout>
-                <button text={"Open yellow modal"} onTap={() => {
-                    const currentPage: Page = forwardedRef.current!.nativeView;
-                    currentPage.showModal(yellowPageRef.current!.nativeView, {
-                        context: {},
-                        closeCallback: () => { },
-                        animated: true,
-                        stretched: false,
-                    });
-                }} />
-            </stackLayout>
-
-            <PortalToPageWithActionBar forwardedRef={yellowPageRef} actionBarTitle={"Yellow page"} backgroundColor={"yellow"}>
+        return (
+            <page ref={forwardedRef} actionBarHidden={false} {...rest}>
+                <actionBar title="Navigation Hub" className="action-bar" />
                 <stackLayout>
-                    <label>You're viewing the yellow page!</label>
-                    <button text={"Open green modal"} onTap={() => {
-                        const currentPage: Page = yellowPageRef.current!.nativeView;
-                        currentPage.showModal(greenPageRef.current!.nativeView, {
-                            context: {},
-                            closeCallback: () => { },
-                            animated: true,
-                            stretched: false,
-                        });
-                    }} />
-                    <button text={"Close yellow modal"} onTap={() => {
-                        const currentPage: Page = yellowPageRef.current!.nativeView;
-                        currentPage.closeModal({});
-                    }} />
+                    <button
+                        text={"Open yellow modal"}
+                        onTap={() => {
+                            const currentPage: Page = forwardedRef.current!.nativeView;
+                            currentPage.showModal(yellowPageRef.current!.nativeView, {
+                                context: {},
+                                closeCallback: () => {},
+                                animated: true,
+                                stretched: false,
+                            });
+                        }}
+                    />
                 </stackLayout>
-            </PortalToPageWithActionBar>
 
-            <PortalToPageWithActionBar forwardedRef={greenPageRef} actionBarTitle={"Green page"} backgroundColor={"green"}>
-                <stackLayout>
-                    <label>You're viewing the green page!</label>
-                    <button text={"Close green modal"} onTap={() => {
-                        const currentPage: Page = greenPageRef.current!.nativeView;
-                        currentPage.closeModal({});
-                    }} />
-                </stackLayout>
-            </PortalToPageWithActionBar>
-        </page>);
+                <PortalToPageWithActionBar forwardedRef={yellowPageRef} actionBarTitle={"Yellow page"} backgroundColor={"yellow"}>
+                    <stackLayout>
+                        <label>You're viewing the yellow page!</label>
+                        <button
+                            text={"Open green modal"}
+                            onTap={() => {
+                                const currentPage: Page = yellowPageRef.current!.nativeView;
+                                currentPage.showModal(greenPageRef.current!.nativeView, {
+                                    context: {},
+                                    closeCallback: () => {},
+                                    animated: true,
+                                    stretched: false,
+                                });
+                            }}
+                        />
+                        <button
+                            text={"Close yellow modal"}
+                            onTap={() => {
+                                const currentPage: Page = yellowPageRef.current!.nativeView;
+                                currentPage.closeModal({});
+                            }}
+                        />
+                    </stackLayout>
+                </PortalToPageWithActionBar>
+
+                <PortalToPageWithActionBar forwardedRef={greenPageRef} actionBarTitle={"Green page"} backgroundColor={"green"}>
+                    <stackLayout>
+                        <label>You're viewing the green page!</label>
+                        <button
+                            text={"Close green modal"}
+                            onTap={() => {
+                                const currentPage: Page = greenPageRef.current!.nativeView;
+                                currentPage.closeModal({});
+                            }}
+                        />
+                    </stackLayout>
+                </PortalToPageWithActionBar>
+            </page>
+        );
     }
 }
 
-
 export class ActionBarTest extends React.Component<{}, {}> {
-    render(){
+    render() {
         const navigationButton = new NavigationButton();
         navigationButton.text = "Go Back";
         return React.createElement(
@@ -119,11 +145,11 @@ export class ActionBarTest extends React.Component<{}, {}> {
 }
 
 export class TabViewTest extends React.Component<{}, {}> {
-    render(){
+    render() {
         return React.createElement(
             "tabView",
             {
-                selectedIndex: 1
+                selectedIndex: 1,
             },
 
             React.createElement(
@@ -131,11 +157,7 @@ export class TabViewTest extends React.Component<{}, {}> {
                 {
                     title: "Dock",
                 },
-                React.createElement(
-                    DockLayoutTest,
-                    {},
-                    null
-                )
+                React.createElement(DockLayoutTest, {}, null)
             ),
 
             React.createElement(
@@ -143,12 +165,8 @@ export class TabViewTest extends React.Component<{}, {}> {
                 {
                     title: "Flexbox",
                 },
-                React.createElement(
-                    FlexboxLayoutTest,
-                    {},
-                    null
-                )
-            ),
+                React.createElement(FlexboxLayoutTest, {}, null)
+            )
 
             // React.createElement(
             //     ReactTabViewItem,
@@ -186,17 +204,19 @@ export class TabViewTest extends React.Component<{}, {}> {
 }
 
 export class PageWithActionBar extends React.Component<
-    {
-        actionBarTitle?: string,
-        forwardedRef: React.RefObject<NSVElement<Page>>,
-    } & PageAttributes,
+    PropsWithChildren<
+        {
+            actionBarTitle?: string;
+            forwardedRef: React.RefObject<NSVElement<Page>>;
+        } & PageAttributes
+    >,
     {}
 > {
-    render(){
+    render() {
         const { children, forwardedRef, actionBarTitle, ...rest } = this.props;
 
         return (
-            <page ref={forwardedRef} actionBarHidden={false} {...rest} >
+            <page ref={forwardedRef} actionBarHidden={false} {...rest}>
                 <actionBar {...{ title: actionBarTitle }} />
                 {children}
             </page>
@@ -204,7 +224,7 @@ export class PageWithActionBar extends React.Component<
     }
 }
 
-export function FrameWithPageWithActionBarNew(props: PropsWithChildren<PageAttributes>){
+export function FrameWithPageWithActionBarNew(props: PropsWithChildren<PageAttributes>) {
     const { children, ...rest } = props;
 
     return (
@@ -217,15 +237,14 @@ export function FrameWithPageWithActionBarNew(props: PropsWithChildren<PageAttri
     );
 }
 
-
-export function PageWithActionBarNew(props: PropsWithChildren<PageAttributes>){
+export function PageWithActionBarNew(props: PropsWithChildren<PageAttributes>) {
     const { children, ...rest } = props;
 
     return (
         <page actionBarHidden={false} {...rest}>
             <actionBar>
                 <label nodeRole={"titleView"}>Hello Title View</label>
-                
+
                 <actionItem nodeRole={"actionItems"}>
                     <button nodeRole={"actionView"}>One</button>
                 </actionItem>
@@ -241,22 +260,21 @@ export function PageWithActionBarNew(props: PropsWithChildren<PageAttributes>){
     );
 }
 
-
 export class PageWithComplexActionBarTest extends React.Component<
-    {
-        actionBarTitle?: string,
-        forwardedRef: React.RefObject<NSVElement<Page>>,
-    },
+    PropsWithChildren<{
+        actionBarTitle?: string;
+        forwardedRef: React.RefObject<NSVElement<Page>>;
+    }>,
     {}
 > {
-    render(){
+    render() {
         const { children, forwardedRef, ...rest } = this.props;
 
         return (
-            <page ref={forwardedRef} actionBarHidden={false} {...rest} >
+            <page ref={forwardedRef} actionBarHidden={false} {...rest}>
                 <actionBar>
                     {/* The Switch will become the titleView */}
-                    <switch/>
+                    <switch />
                     {/* The ActionItem will be added to the actionItems array */}
                     <actionItem text={"AI"} ios={{ position: "right" as const, systemIcon: 4 }}></actionItem>
                     {/* The NavigationButton will set as the NavigationButton (but won't be visible because there's no backwards navigation to do from here). */}
@@ -268,44 +286,48 @@ export class PageWithComplexActionBarTest extends React.Component<
     }
 }
 
-export class FramedPageWithComplexActionBarTest extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Frame>> }, {}> {
+export class FramedPageWithComplexActionBarTest extends React.Component<
+    PropsWithChildren<{ forwardedRef: React.RefObject<NSVElement<Frame>> }>,
+    {}
+> {
     private readonly pageWithActionBarRef = React.createRef<NSVElement<Page>>();
 
-    render(){
+    render() {
         const { forwardedRef, children, ...rest } = this.props;
 
         return (
-            <FramedChildTest forwardedRef={forwardedRef} childPageRef={this.pageWithActionBarRef} {...rest} >
-                <PageWithComplexActionBarTest forwardedRef={this.pageWithActionBarRef}>
-                    {children}
-                </PageWithComplexActionBarTest>
+            <FramedChildTest forwardedRef={forwardedRef} childPageRef={this.pageWithActionBarRef} {...rest}>
+                <PageWithComplexActionBarTest forwardedRef={this.pageWithActionBarRef}>{children}</PageWithComplexActionBarTest>
             </FramedChildTest>
         );
     }
 }
 
-const PortalToPage: React.SFC<{ forwardedRef: React.RefObject<NSVElement<Page>>, actionBarTitle: string }> = (props) => {
+const PortalToPage: React.FC<
+    PropsWithChildren<{
+        forwardedRef: React.RefObject<NSVElement<Page>>;
+        actionBarTitle: string;
+    }>
+> = (props) => {
     const { forwardedRef, actionBarTitle, children } = props;
     return ReactNativeScript.createPortal(
-        (
-            <PageWithActionBar forwardedRef={forwardedRef} actionBarTitle={actionBarTitle}>
-                {children}
-            </PageWithActionBar>
-        ),
+        <PageWithActionBar forwardedRef={forwardedRef} actionBarTitle={actionBarTitle}>
+            {children}
+        </PageWithActionBar>,
         null,
         `Portal('${actionBarTitle}')`
     );
-}
+};
 
 export class HubTest extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Page>> }, {}> {
     private readonly absoluteLayoutPageRef = React.createRef<NSVElement<Page>>();
     private readonly dockLayoutPageRef = React.createRef<NSVElement<Page>>();
     private readonly flexboxLayoutPageRef = React.createRef<NSVElement<Page>>();
 
-    private navigateToPage(targetPage: Page, title: string){
+    private navigateToPage(targetPage: Page, title: string) {
         const page: Page = this.props.forwardedRef.current!.nativeView;
-        const frame: Frame|undefined = page.frame;
-        if(!frame){
+        const frame: Frame | undefined = page.frame;
+        if (!frame) {
             console.error(`No frame found for page ${page}. Ensure that HubTest is embedded in a Frame (e.g. via FramedHubTest).`);
             return;
         }
@@ -313,11 +335,11 @@ export class HubTest extends React.Component<{ forwardedRef: React.RefObject<NSV
             create: () => {
                 console.log(`Navigating from ${page} to ${title} page. Ref:`, targetPage);
                 return targetPage;
-            }
+            },
         });
     }
 
-    render(){
+    render() {
         const { forwardedRef } = this.props;
 
         return (
@@ -343,17 +365,17 @@ export class HubTest extends React.Component<{ forwardedRef: React.RefObject<NSV
                         }}
                     />
                 </stackLayout>
-                
+
                 <PortalToPage forwardedRef={this.absoluteLayoutPageRef} actionBarTitle={"AbsoluteLayout"}>
-                    <AbsoluteLayoutTest/>
+                    <AbsoluteLayoutTest />
                 </PortalToPage>
-                
+
                 <PortalToPage forwardedRef={this.dockLayoutPageRef} actionBarTitle={"DockLayout"}>
-                    <DockLayoutTest/>
+                    <DockLayoutTest />
                 </PortalToPage>
-                
+
                 <PortalToPage forwardedRef={this.flexboxLayoutPageRef} actionBarTitle={"FlexboxLayout"}>
-                    <FlexboxLayoutTest/>
+                    <FlexboxLayoutTest />
                 </PortalToPage>
             </page>
         );
@@ -362,45 +384,51 @@ export class HubTest extends React.Component<{ forwardedRef: React.RefObject<NSV
 
 /**
  * MUST render the portal into a container (but only via a Host Config hack, as Pages can't have containers) or null.
- * 
+ *
  * Rendering into forwardedRef.current mysteriously works on the first render, but only because forwardedRef.current
  * is null at first. On second render (due to a state change of any component within), forwardedRef.current will become
  * populated and thus cause this to break (because it's not a container of the root component; it IS the root component).
- * 
+ *
  * ... So we render into null. This seems to work perfectly, but we'll probably have to add special-case handling in the
  * Host Config so that if the root component (Page) is ever unmounted, we don't call null.removeChild().
  *
  * The converse case of outerparent.removeChild(null) (when we unmount the Portal) might not need special handling,
  * because null is not regarded as a ReactNode to begin with.
  */
-export const PortalToPageWithActionBar: React.SFC<
-    { forwardedRef: React.RefObject<NSVElement<Page>>, actionBarTitle: string } & PageAttributes
-> =
-(props) => {
+export const PortalToPageWithActionBar: React.FC<
+    PropsWithChildren<
+        {
+            forwardedRef: React.RefObject<NSVElement<Page>>;
+            actionBarTitle: string;
+        } & PageAttributes
+    >
+> = (props) => {
     const { forwardedRef, actionBarTitle, children, ...rest } = props;
     console.log(`[PortalToPageWithActionBar - "${actionBarTitle}"] createPortal() forwardedRef.current: ${forwardedRef.current}`);
     return ReactNativeScript.createPortal(
-        (
-            <page ref={forwardedRef} actionBarHidden={false} {...rest} >
-                <actionBar title={actionBarTitle} className={"action-bar"}/>
-                {children}
-            </page>
-        ),
+        <page ref={forwardedRef} actionBarHidden={false} {...rest}>
+            <actionBar title={actionBarTitle} className={"action-bar"} />
+            {children}
+        </page>,
         null,
         `Portal('${actionBarTitle}')`
     );
-}
+};
 
 /**
  * Above, we use a Stateless Functional Component. Here is the equivalent using a regular class component.
- * 
- * We may find that a class component is necessary to provide the lifecycle methods to clean up upon unmount 
+ *
+ * We may find that a class component is necessary to provide the lifecycle methods to clean up upon unmount
  * (componentWillUnmount), but also... maybe not.
- * 
+ *
  * An explicit shouldComponentUpdate() is purely there to help me follow the logs. It's not needed otherwise.
  */
 export class StatefulPortalToPageWithActionBar extends React.Component<
-    { forwardedRef: React.RefObject<NSVElement<Page>>, actionBarTitle: string } & PageAttributes,
+    PropsWithChildren<{
+        forwardedRef: React.RefObject<NSVElement<Page>>;
+        actionBarTitle: string;
+    }> &
+        PageAttributes,
     {}
 > {
     shouldComponentUpdate(
@@ -411,17 +439,17 @@ export class StatefulPortalToPageWithActionBar extends React.Component<
         return true;
     }
 
-    render(){
+    render() {
         const { forwardedRef, actionBarTitle, children, ...rest } = this.props;
-        console.log(`[StatefulPortalToPageWithActionBar - "${actionBarTitle}"] createPortal() forwardedRef.current: ${forwardedRef.current}`);
+        console.log(
+            `[StatefulPortalToPageWithActionBar - "${actionBarTitle}"] createPortal() forwardedRef.current: ${forwardedRef.current}`
+        );
 
         return ReactNativeScript.createPortal(
-            (
-                <page actionBarHidden={false} {...rest} ref={forwardedRef} >
-                    <actionBar title={actionBarTitle} className={"action-bar"}/>
-                    {children}
-                </page>
-            ),
+            <page actionBarHidden={false} {...rest} ref={forwardedRef}>
+                <actionBar title={actionBarTitle} className={"action-bar"} />
+                {children}
+            </page>,
             null,
             `Portal('${actionBarTitle}')`
         );
@@ -430,7 +458,7 @@ export class StatefulPortalToPageWithActionBar extends React.Component<
 
 export class SimpleHub extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Page>> } & PageAttributes, {}> {
     private readonly bluePageRef = React.createRef<NSVElement<Page>>();
-    render(){
+    render() {
         const { forwardedRef, ...rest } = this.props;
 
         return (
@@ -444,12 +472,12 @@ export class SimpleHub extends React.Component<{ forwardedRef: React.RefObject<N
                             currentPage.frame.navigate({
                                 create: () => {
                                     return this.bluePageRef.current.nativeView;
-                                }
+                                },
                             });
                         }}
                     />
                 </stackLayout>
-                
+
                 <PortalToPageWithActionBar forwardedRef={this.bluePageRef} actionBarTitle={"Blue page"} backgroundColor={"blue"}>
                     <label>You're viewing the blue page!</label>
                 </PortalToPageWithActionBar>
@@ -461,20 +489,20 @@ export class SimpleHub extends React.Component<{ forwardedRef: React.RefObject<N
 export class FrameTest extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Frame>> } & FrameAttributes, {}> {
     private readonly bluePageRef = React.createRef<NSVElement<Page>>();
 
-    componentDidMount(){
-        const node: Frame|null = this.props.forwardedRef.current.nativeView;
-        if(node){
+    componentDidMount() {
+        const node: Frame | null = this.props.forwardedRef.current.nativeView;
+        if (node) {
             node.navigate({
                 create: () => {
                     return this.bluePageRef.current.nativeView;
-                }
+                },
             });
         } else {
             console.warn(`React ref to NativeScript View lost, so unable to update event listeners.`);
         }
     }
 
-    render(){
+    render() {
         const { forwardedRef, ...rest } = this.props;
 
         return (
@@ -488,16 +516,15 @@ export class FrameTest extends React.Component<{ forwardedRef: React.RefObject<N
 }
 
 export class FramedChildTest extends React.Component<
-    {
-        forwardedRef: React.RefObject<NSVElement<Frame>>,
-        childPageRef: React.RefObject<NSVElement<Page>>,
-    },
+    PropsWithChildren<{
+        forwardedRef: React.RefObject<NSVElement<Frame>>;
+        childPageRef: React.RefObject<NSVElement<Page>>;
+    }>,
     {}
->
-{
-    componentDidMount(){
-        const node: Frame|null = this.props.forwardedRef.current.nativeView;
-        if(!node){
+> {
+    componentDidMount() {
+        const node: Frame | null = this.props.forwardedRef.current.nativeView;
+        if (!node) {
             console.warn(`[FramedChildTest] React ref to NativeScript View lost, so unable to update event listeners.`);
             return;
         }
@@ -505,43 +532,35 @@ export class FramedChildTest extends React.Component<
         console.log(`[FramedChildTest] componentDidMount; shall navigate to page within.`);
         node.navigate({
             create: () => {
-                const childPage: Page|undefined = this.props.childPageRef.current!.nativeView;
+                const childPage: Page | undefined = this.props.childPageRef.current!.nativeView;
                 console.log(`[FramedChildTest] create(); shall return ref to page: ${childPage}`);
                 return childPage;
-            }
+            },
         });
     }
 
-    componentWillUnmount(){
+    componentWillUnmount() {
         console.log(`[FramedChildTest] componentWillUnmount`);
     }
 
-    render(){
+    render() {
         return (
             <frame ref={this.props.forwardedRef}>
-                {(
-                    ReactNativeScript.createPortal(
-                        (
-                            this.props.children
-                        ),
-                        null,
-                        `Portal('FramedChild')`
-                    )
-                )}
+                {ReactNativeScript.createPortal(this.props.children, null, `Portal('FramedChild')`)}
             </frame>
         );
     }
 }
 
-export class FramedHubTest extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Frame>> }, {}> {
+export class FramedHubTest extends React.Component<PropsWithChildren<{ forwardedRef: React.RefObject<NSVElement<Frame>> }>, {}> {
     private readonly hubTestPageRef = React.createRef<NSVElement<Page>>();
 
-    render(){
+    render() {
         const { forwardedRef, children, ...rest } = this.props;
 
         return (
-            <FramedChildTest forwardedRef={forwardedRef} childPageRef={this.hubTestPageRef} {...rest} >
-                <HubTest forwardedRef={this.hubTestPageRef}/>
+            <FramedChildTest forwardedRef={forwardedRef} childPageRef={this.hubTestPageRef} {...rest}>
+                <HubTest forwardedRef={this.hubTestPageRef} />
             </FramedChildTest>
         );
     }
@@ -550,32 +569,28 @@ export class FramedHubTest extends React.Component<{ forwardedRef: React.RefObje
 export class FramedLayoutTest extends React.Component<{ forwardedRef: React.RefObject<NSVElement<Frame>> }, {}> {
     private readonly layoutTestPageRef = React.createRef<NSVElement<Page>>();
 
-    componentDidMount(){
-        const node: Frame|null = this.props.forwardedRef.current.nativeView;
-        if(node){
+    componentDidMount() {
+        const node: Frame | null = this.props.forwardedRef.current.nativeView;
+        if (node) {
             node.navigate({
                 create: () => {
                     return this.layoutTestPageRef.current.nativeView;
-                }
+                },
             });
         } else {
             console.warn(`React ref to NativeScript View lost, so unable to update event listeners.`);
         }
     }
 
-    render(){
+    render() {
         return (
             <frame ref={this.props.forwardedRef}>
-                {(
-                    ReactNativeScript.createPortal(
-                        (
-                            <contentView ref={this.layoutTestPageRef}>
-                                <DockLayoutTest/>
-                            </contentView>
-                        ),
-                        null,
-                        `Portal('Dock Layout Test')`
-                    )
+                {ReactNativeScript.createPortal(
+                    <contentView ref={this.layoutTestPageRef}>
+                        <DockLayoutTest />
+                    </contentView>,
+                    null,
+                    `Portal('Dock Layout Test')`
                 )}
             </frame>
         );
