+++ sample/src/testComponents/propAlteration.tsx
@@ -3,10 +3,8 @@ import { GestureEventData } from "@nativescript/core";
 
 export function UpdateEventHandlerTest() {
     function handleTap(args: GestureEventData) {
-      console.log('2');
+        console.log("2");
     }
-  
-    return (
-        <button onTap={handleTap} text="click me"/>
-    );
+
+    return <button onTap={handleTap} text="click me" />;
 }
