--- sample/tsfmt.json
@@ -1,4 +1,0 @@
-{
-  "indentSize": 4,
-  "tabSize": 4
-}
